add 2020 CommunityController with routes and move searchTargetById
