+++ api/app/Http/Controllers/Year2019/VolunteerController.php
@@ -24,28 +24,4 @@ class VolunteerController extends BaseVolunteerController
         return $this->returnSuccess('success', $result);
     }
 
-    /**
-     * 由 id 取出特定社群
-     *
-     * @param array $dataset
-     * @param integer $targetId
-     * @return array
-     */
-    private function searchTargetById($dataset, $targetId)
-    {
-        $result = array_filter($dataset, function ($subset) use ($targetId) {
-            if ((int)$subset['id'] === (int)$targetId) {
-                return true;
-            }
-        });
-        $result = array_pop($result);
-        if (isset($result['photo']) && $result['photo'] !== '') {
-            $result['photo'] = url($result['photo']);
-        }
-        if (isset($result['introduction_en']) && $result['introduction_en'] === '') {
-            $result['introduction_en'] = $result['introduction'];
-        }
-        unset($result['id']);
-        return $result;
-    }
 }
