+++ api/routes/web.php
@@ -74,6 +74,12 @@ $router->group(['prefix' => 'api'], function ($router) {
                 $router->get('/{id}', 'VolunteerController@show');
                 $router->get('/images/{name}', 'VolunteerController@imagesView');
             });
+            $router->group(['prefix' => 'community'], function ($router) {
+                $router->get('/', 'CommunityController@index');
+                $router->get('/organizer/{id}', 'CommunityController@getOrganizer');
+                $router->get('/participant/{id}', 'CommunityController@getParticipant');
+                $router->get('/images/{name}', 'CommunityController@imagesView');
+            });
         });
     });
 });
